requirements.txt
alabaster==0.7.8
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.12.12
asttokens==2.2.1
attrs==22.1.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==5.0.1
blinker==1.4
bottle==0.12.23
breezy==3.1.0
Brlapi==0.8.2
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.1.3
cmake==3.26.1
cmdtest==0.32+git
Coffin==2.0.1
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.5
cryptography==39.0.0
cupshelpers==1.0
cycler==0.11.0
databind==1.5.3
databind.core==1.5.3
databind.json==1.5.3
dbus-python==1.2.16
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dirsearch==0.4.3.post1
distlib==0.3.6
distro==1.5.0
dj-database-url==1.0.0
Django==4.2
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-filter==2.4.0
django-render==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
docspec==2.0.2
docspec-python==2.0.2
docstring-parser==0.11
docutils==0.16
dulwich==0.20.15
emoji==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastimport==0.9.8
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.4
fqdn==1.5.1
galternatives==1.0.8
gpg===1.14.0-unknown
greenlet==0.4.17
gunicorn==20.1.0
html5lib==1.1
httpie==3.2.1
httplib2==0.18.1
identify==2.5.8
idna==2.10
imagesize==1.2.0
importlib-metadata==5.0.0
instaloader==4.9.5
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
isc==2.0
isoduration==20.11.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.6.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==7.4.1
jupyter-core==5.1.3
jupyter-events==0.6.0
jupyter-server==1.23.4
jupyter-server-terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.18.0
kazam==1.4.5
keyring==22.0.1
kiwisolver==1.4.4
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.8.0
lit==16.0.0
llvmlite==0.39.1
louis==3.16.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mercurial==5.6.1
meson==0.56.2
meteo-qt==2.1
mistune==2.0.4
mpmath==1.2.1
mps-youtube==0.2.8
msgpack==1.0.0
multidict==6.0.4
mysql-connector-python==8.0.32
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8
nodeenv==1.7.0
notebook==6.5.2
notebook-shim==0.2.2
nr.util==0.8.12
ntlm-auth==1.5.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
OpenTimelineIO==0.15.0
packaging==21.3
pafy==0.5.5
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patiencediff==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.3
ply==3.11
pockets==0.9.1
pre-commit==2.20.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.8.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaaf2==1.6.0
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pycurl==7.43.0.6
pydoc-markdown==4.6.4
pydub==0.25.1
PyGithub==1.43.7
Pygments==2.7.1
PyGObject==3.38.0
PyJWT==1.7.1
pylint==2.15.5
pynvim==0.4.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.19.3
PySimpleSOAP==1.16.2
pysmbc==1.0.23
PySocks==1.7.1
pytest==3.0.5
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-gitlab==2.5.0
python-json-logger==2.0.4
python-xlib==0.31
pytz==2021.1
PyX==0.16
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
ranger-fm==1.9.3
regex==2022.10.31
render==1.0.0
render-python==2.2.2
reportbug==7.10.3+deb11u1
reportlab==3.5.59
requests==2.28.2
requests-futures==1.0.0
requests-ntlm==1.1.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.0.0
roman==2.0.0
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.97
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.1.0
soupsieve==2.2.1
SpeechRecognition==3.9.0
Sphinx==3.4.3
sphinxcontrib-napoleon==0.7
sqlparse==0.4.1
stack-data==0.6.2
stem==1.8.1
sympy==1.11.1
tabulate==0.9.0
termcolor==2.1.1
terminado==0.17.1
terminator==2.1.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomli-w==1.0.0
tomlkit==0.11.6
torch==2.0.0
tornado==6.2
torrequest==0.1.0
tqdm==4.64.1
traitlets==5.8.1
triton==2.0.0
ttystatus==0.38
typer==0.7.0
typing-extensions==4.4.0
ueberzug==18.1.9
ufw==0.36
uri-template==1.2.0
urllib3==1.26.5
virtualenv==20.16.6
voila==0.4.0
wadllib==1.3.5
watchdog==2.2.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
wifite==2.5.8
WoeUSB-ng==0.2.10.post2
wordcloud==1.8.2.2
wrapt==1.14.1
wxPython==4.0.7
xdg==5
yapf==0.32.0
yarg==0.1.9
youtube-dl==2021.6.6
zipp==3.10.0
